# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: payd HTTP API
  version: 1.0.0
  description: |
    The HTTP API served by payd, a self-hosted TRON payment processor. It creates
    payment orders against pooled TRON addresses, reports the payments detected on
    chain for them, and moves funds out again through operator-approved withdrawals.

    **Money is always a decimal string**, never a JSON number — `"1.5"`, not `1.5`.
    JSON numbers cannot represent every base-unit value exactly, so every amount
    field in every request and response is a string. Fields ending in `_raw` are in
    base units (SUN for TRX, 6-decimal units for USDT); all other amount fields are
    in whole asset units.

    **Authentication** is a shared secret in the `X-API-Key` request header. Each key
    carries its own set of scopes (`orders:read`, `orders:write`, `wallets:read`,
    `wallets:write`, `withdrawals:read`, `withdrawals:write`, `resources:write`,
    `admin:read`), and every operation
    below states the one scope it requires. A missing key, a wrong key, or a key
    without the required scope all return the same 401 with no further detail.

    **Rate limits** are counted per API key in fixed one-minute windows: 100 requests
    per minute in general, and 10 requests per minute on `/api/v1/withdrawals` and its
    subpaths. Exceeding either returns 429 `rate_limited`.

    **Errors** always use the envelope `{"error": {"code", "message", "details"}}`.
...

...
servers:
  - url: http://127.0.0.1:8080
security:
  - apiKey: []
...
paths:
...
  /api/v1/withdrawals:
    post:
      tags: [Withdrawals]
      summary: Create or replay a withdrawal request
      description: |
        Requests a transfer of funds out of a pooled address to an external address.

        Requires scope `withdrawals:write`. Rate limit **10 requests/minute** per API key
        on this path and its subpaths, not the usual 100.

        **This endpoint only records the request.** It returns as soon as the row is
        stored in status `requested`; a background worker acquires resources, signs and
        broadcasts the transaction afterwards. Poll `GET /api/v1/withdrawals/{id}` for the
        outcome. A 201 means "accepted", never "sent".

        **Idempotency-Key is mandatory** and is checked *before* the TOTP code is even
        looked at. That ordering is deliberate: a client that times out and retries sends
        the same key together with the same TOTP code, and treating that as a replayed
        code would reject an honest retry. A missing header is 400
        `missing_idempotency_key`.

        - Key already used, and `from_address`, `to_address`, `asset` and `amount` all
          match: 200 with the stored withdrawal. Nothing new is created and the TOTP code
          is not consumed a second time.
        - Key already used with any different value: 409 `idempotency_key_reuse`. Pick a
          fresh key for a genuinely different transfer.
        - New key: 201 with the newly created withdrawal.

        **TOTP.** When required by configuration, send a 6-digit code from the configured
        authenticator secret in the **`X-TOTP` header** — never in the request body, which
        is rejected with 400 `totp_in_body`. Codes are accepted within one 30-second step
        either side of the current one, and each code is consumed once — its use is stored
        in the database, so a restart does not make it reusable. Any failure (bad code,
        already used, no secret configured) is 401 `invalid_totp`. Neither headers nor
        bodies are written to the logs, so codes do not leak there. If a new request is
        rejected with 409 after successful TOTP validation, `details.totp_consumed` is
        `true`; wait for a fresh code before submitting a corrected request.

        **Checks, in the order they are applied.** Withdrawals disabled in configuration:
        503 `withdrawals_disabled`. Unknown asset: 400 `invalid_asset`. Malformed
        destination address or a non-positive amount: 400 `invalid_withdrawal`. No fresh
        price quote for the USD valuation: 503 `price_unavailable`. Source address unknown
        or disabled: 400 `invalid_source`. Source flagged for balance drift: 409
        `balance_drift` — clear it via `POST /api/v1/wallets/{address}/clear-drift` first.
        Amount above the source's **confirmed** balance: 409
        `insufficient_confirmed_balance`; pending funds are never spendable. Amount that
        would push the running USD total past the daily cap: 409 `daily_limit_exceeded` —
        check the remaining allowance with `GET /api/v1/withdrawals/limits`. The cap is
        enforced as part of the same database write, so two concurrent requests cannot
        both slip under it.

        **There is no retry, resume or re-broadcast endpoint anywhere in this API.** A
        withdrawal that fails is terminal; to try again, submit a new request with a new
        `Idempotency-Key`. This is a deliberate safety property — nothing in the system can
        re-broadcast a signed transfer, so a stuck withdrawal can never quietly send twice.

        **Side effects.** Stores the withdrawal in status `requested`, recording the API
        key name and the caller's IP as the requester, and writes an audit-log entry.
      x-required-scope: withdrawals:write
      x-requirements: [API-003, API-015, API-020, API-022, API-023, API-024, API-026, WDR-001a]
      x-error-codes:
        400: [invalid_request, missing_idempotency_key, totp_in_body, invalid_asset, invalid_withdrawal, invalid_source]
        401: [unauthorized, invalid_totp]
        409: [idempotency_key_reuse, balance_drift, daily_limit_exceeded, insufficient_confirmed_balance]
        429: [rate_limited]
        500: [internal_error]
        503: [withdrawals_disabled, price_unavailable]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: >-
            Caller-chosen key that identifies this withdrawal request. Reuse it verbatim
            when retrying a request whose response you did not receive; use a new one for
            every genuinely new transfer.
          schema: {type: string, example: w-001}
        - name: X-TOTP
          in: header
          required: false
          description: >-
            Six-digit authenticator code, required when `withdrawal.require_totp` is set.
            Sent as a header so the code never enters a request body that a proxy or client
            might log or replay. A code placed in the body instead is rejected with 400
            `totp_in_body`.
          schema: {type: string, pattern: '^[0-9]{6}$', example: '482910'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateWithdrawalRequest'}
      responses:
        '200':
          description: >-
            Replay. The Idempotency-Key matched an existing request with identical
            details, so that withdrawal is returned as-is; no second transfer was created
            and the TOTP code was not consumed again.
          content: {application/json: {schema: {$ref: '#/components/schemas/Withdrawal'}}}
        '201':
          description: >-
            Accepted and stored in status `requested`. Signing and broadcast happen
            asynchronously; poll `GET /api/v1/withdrawals/{id}` for the outcome.
          content: {application/json: {schema: {$ref: '#/components/schemas/Withdrawal'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      tags: [Withdrawals]
      summary: List withdrawals
      description: |
        Lists withdrawal requests, newest first, optionally filtered to one exact `status`.

        Requires scope `withdrawals:read`. Rate limit 10 requests/minute per API key, as
        on every `/api/v1/withdrawals` path.

        The response envelope is `{"items": [...], "next_cursor": "..."}`. Pass a
        non-empty `next_cursor` back unchanged as `cursor` to continue through the
        newest-first result set; an empty value marks the final page. `limit` is 1-200
        and defaults to 50.
      x-required-scope: withdrawals:read
      x-requirements: [API-003, API-020, API-023, API-024, API-025]
      x-error-codes:
        400: [invalid_pagination]
        401: [unauthorized]
        429: [rate_limited]
        500: [internal_error]
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - name: status
          in: query
          schema: {type: string, example: requested}
      responses:
        '200':
          description: Withdrawals, newest first, with an opaque next-page cursor.
          content: {application/json: {schema: {$ref: '#/components/schemas/WithdrawalList'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Send the API key itself in this header — the plaintext secret, not the Argon2id
        `key_hash` that appears in the server's configuration file.

        The server stores only the hash and verifies every request by recomputing it,
        comparing in constant time. It never learns the key from configuration, and a
        leaked configuration file does not yield usable keys.

        Each key carries its own scopes. A missing header, an unknown key, and a valid key
        lacking the scope an operation requires all return the same 401 `unauthorized`
        with no detail about which it was — that is deliberate, so the response cannot be
        used to probe for valid keys.
  parameters:
    Limit:
      name: limit
      in: query
      description: >-
        Maximum rows to return in this page. Values outside 1-200 are rejected with
        400 `invalid_pagination`.
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    Cursor:
      name: cursor
      in: query
      description: >-
        Position to resume from. Pass back the `next_cursor` returned by the previous
        page, unmodified; omit it for the first page. Treat the value as opaque — its
        encoding is not part of the contract. Anything that is not a cursor this API
        issued is 400 `invalid_pagination`.
      schema: {type: string}
...
  responses:
    BadRequest:
      description: >-
        The request itself is wrong: malformed JSON, an unknown JSON field, a body over
        1 MiB, a missing required field, or a value that fails validation. Fix the request
        before retrying — repeating it unchanged gives the same result. The specific
        `code` says which; each operation lists the ones it can return.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unauthorized:
      description: >-
        The caller is not allowed to make this call. Returned for a missing or wrong
        `X-API-Key`, for a valid key that lacks the scope this operation requires, and —
        on withdrawal creation — for a TOTP code that is wrong or already used. The first
        two are indistinguishable by design and return `unauthorized`; a TOTP failure
        returns `invalid_totp`.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    Conflict:
      description: >-
        The request is well-formed but disagrees with stored state — a reused idempotency
        key with different values, an order that already received funds, a source address
        flagged for drift, or a limit that would be exceeded. Read the `code` and act on
        it; retrying the same request unchanged will not succeed.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    RateLimited:
      description: >-
        Too many requests from this API key inside the current one-minute window: 100 per
        minute in general, 10 per minute on `/api/v1/withdrawals` and its subpaths. Windows
        are fixed, not sliding, and counted per key, so other keys are unaffected. Wait for
        the next minute and retry.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    InternalError:
      description: >-
        The processor failed unexpectedly. Nothing about the request needs changing; check
        the server logs. Because writes are transactional, a request that failed this way
        was either applied completely or not at all.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unavailable:
      description: >-
        The request is fine but the processor cannot serve it right now — no fresh price
        quote, withdrawals turned off in configuration, or the address pool exhausted.
        These are transient states worth retrying later, unlike a 400 or a 409.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code:
              type: string
              description: >-
                Stable machine-readable identifier for what went wrong. Branch on this.
                The set of codes an operation can return is listed under its
                `x-error-codes` extension.
              example: insufficient_balance
            message:
              type: string
              description: >-
                Human-readable explanation, for logs and operator consoles. Wording may
                change between releases; never match on it.
              example: confirmed balance is insufficient
            details:
              type: object
              additionalProperties: true
              description: >-
                Extra machine-readable context for codes that have any, otherwise empty.
                For `external_ref_conflict` it carries `fields`, the list of order
                attributes that disagreed. A withdrawal-creation 409 after successful
                TOTP validation carries `totp_consumed: true`.
              example: {}
...
    CreateWithdrawalRequest:
      type: object
      additionalProperties: false
      required: [from_address, to_address, asset, amount]
      properties:
        from_address:
          type: string
          description: >-
            Pooled address to send from. It must be known to the processor, not disabled,
            not flagged for balance drift, and hold enough **confirmed** balance.
          example: TUEZSdKsoDHQMeZwihtdoBiN46zxhGWYdH
        to_address:
          type: string
          description: >-
            External TRON address to send to. Validated for format only — the processor
            cannot tell you whether it is the right destination, and a transfer cannot be
            recalled.
          example: TVjsyZ7fYF3qLF6BQgPmTEZy1xrNNyVAAA
        asset:
          type: string
          description: Symbol of a configured asset to send.
          example: USDT
        amount:
          type: string
          description: Amount to send, in whole asset units, as a positive decimal string.
          example: '1.0'
        totp:
          type: string
          deprecated: true
          writeOnly: true
          description: >-
            **Removed — send the code in the `X-TOTP` header instead.** A request carrying
            this field is rejected with 400 `totp_in_body` rather than being accepted with
            the field ignored, so a caller cannot believe it presented a second factor when
            it did not. The rejected request does not consume the code.
    Withdrawal:
      type: object
      additionalProperties: false
      required: [id, from_address, to_address, asset, amount, amount_raw, amount_usd, status, txid, failure_reason, resolved_by, broadcast_response, fee_raw, network_fee_trx, energy_used, energy_source, energy_cost_trx, bandwidth_source, bandwidth_cost_trx, resource_fee_trx, total_cost_trx, last_lookup_error, created_at, broadcast_at, confirmed_at]
      properties:
        id:
          type: string
          description: Withdrawal identifier, a ULID assigned by the processor.
          example: 01KZK8B5H1KM6XACMG7MK57J8W
        from_address:
          type: string
          description: Pooled address the funds leave.
          example: TUEZSdKsoDHQMeZwihtdoBiN46zxhGWYdH
        to_address:
          type: string
          description: External address the funds go to.
          example: TVjsyZ7fYF3qLF6BQgPmTEZy1xrNNyVAAA
        asset:
          type: string
          description: Symbol of the asset being sent.
          example: USDT
        amount:
          type: string
          description: Amount to send, in whole asset units, as a decimal string.
          example: '1'
        amount_raw:
          type: string
          description: >-
            The same amount in base units (SUN for TRX, 6-decimal units for USDT), as a
            decimal string. This is the exact integer the signed transaction carries.
          example: '1000000'
        amount_usd:
          type: string
          description: >-
            USD value fixed at the moment the request was accepted, as a decimal string.
            This is what counts against the daily cap, so the cap does not move with the
            market.
          example: '1'
        status:
          type: string
          description: >-
            `requested` — accepted, nothing signed yet. `awaiting_resources` /
            `awaiting_energy` — waiting for bandwidth or energy on the source address.
            `signing` — being signed. `broadcast` — sent to the network, awaiting inclusion.
            `confirmed` — final; the funds have left. `failed` — did not go out, see
            `failure_reason`. `needs_operator` — the outcome could not be determined
            automatically and a human must check the chain; this is the one status that
            always warrants an alert.
          example: requested
        txid:
          type: string
          description: >-
            Hash of the broadcast transaction, for explorer lookups. Empty until broadcast.
          example: ''
        failure_reason:
          type: string
          description: Why the withdrawal did not go out. Empty unless it failed.
          example: ''
        resolved_by:
          type: string
          description: >-
            Who closed out a withdrawal that needed a human. Empty unless an operator acted.
          example: ''
        broadcast_response:
          type: string
          description: >-
            The node's raw reply to the broadcast, kept verbatim for diagnosis when the
            transfer was rejected. Empty otherwise.
          example: ''
        fee_raw:
          type: string
          description: >-
            Fee the chain charged, in SUN, as a decimal string. Already includes any TRX
            burned for energy or bandwidth.
          example: '0'
        network_fee_trx:
          type: string
          description: The same chain fee expressed in TRX, as a decimal string.
          example: '0'
        energy_used:
          type: integer
          format: int64
          description: Energy the transaction consumed. Token transfers use energy; plain TRX transfers do not.
          example: 0
        energy_source:
          type: string
          description: >-
            Where the energy came from — the account's own staked energy, a delegation, a
            rental, or TRX burned to cover it. Empty until execution.
          example: ''
        energy_cost_trx:
          type: string
          description: TRX spent on energy, as a decimal string.
          example: '0'
        bandwidth_source:
          type: string
          description: >-
            Where the bandwidth came from — free daily allowance, staked bandwidth, a
            delegation, or TRX burned to cover it. Empty until execution.
          example: ''
        bandwidth_cost_trx:
          type: string
          description: TRX spent on bandwidth, as a decimal string.
          example: '0'
        resource_fee_trx:
          type: string
          description: Energy and bandwidth cost combined, in TRX, as a decimal string.
          example: '0'
        total_cost_trx:
          type: string
          description: >-
            Everything this withdrawal cost in TRX — chain fee plus any rented resources —
            as a decimal string. Burned energy and bandwidth are already inside the chain
            fee and are not counted twice.
          example: '0'
        last_lookup_error:
          type: string
          description: >-
            Why the most recent attempt to check this transaction on chain failed. A
            populated value here means the status may be behind reality, not that the
            withdrawal itself failed.
          example: ''
        created_at:
          type: integer
          format: int64
          description: Unix seconds when the request was accepted.
          example: 1786279000
        broadcast_at:
          type: [integer, 'null']
          format: int64
          description: Unix seconds when the transaction was sent to the network; null until then.
        confirmed_at:
          type: [integer, 'null']
          format: int64
          description: Unix seconds when it became irreversible; null until then.
...
    WithdrawalList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Withdrawal'}
          description: Withdrawals, newest first.
        next_cursor:
          type: string
          description: >-
            Pass back as `cursor` for the next page. Empty means this was the last page.
          example: ''
...